+++ packages/foundry_annotations/lib/foundry_annotations.dart
@@ -1,5 +1,5 @@
 export 'view_state.dart';
-export 'model_state.dart';
+export 'service_state.dart';
 export 'view_model.dart';
-export 'model.dart';
+export 'service.dart';
 export 'view.dart';