+++ packages/foundry_core/lib/foundry_core.dart
@@ -1,7 +1,7 @@
 export 'state_emitter.dart';
 export 'initializable.dart';
 export 'view_model_base.dart';
-export 'stateful_model_base.dart';
+export 'stateful_service_base.dart';
 export 'view_base.dart';
 
 export 'di/scope.dart';