+++ packages/foundry_annotations/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Initial release of foundry_annotations package
 - Added @FoundryViewState annotation for immutable view state
-- Added @FoundryModelState annotation for model state classes
+- Added @FoundryServiceState annotation for service state classes
 - Added @FoundryViewModel annotation for ViewModel classes  
-- Added @FoundryModel annotation for injectable models and services
+- Added @FoundryService annotation for injectable models and services
 - Added @FoundryView annotation for Flutter widgets that bind to ViewModels